use explicit casts

+++ src/ob/timer.hh
@@ -149,7 +149,7 @@ public:
       update();
     }
 
-    long int const now {std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count()};
+    long int const now {static_cast<long int>(std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count())};
 
     return seconds_to_hms(sec - now);
   }
@@ -161,7 +161,7 @@ public:
       update();
     }
 
-    return std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count();
+    return static_cast<long int>(std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count());
   }
 
   template<typename T>
@@ -182,7 +182,7 @@ public:
       update();
     }
 
-    long int const diff {std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count()};
+    long int const diff {static_cast<long int>(std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count())};
 
     return seconds_to_hms(diff);
   }
@@ -194,7 +194,7 @@ public:
       update();
     }
 
-    long int const diff {std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count()};
+    long int const diff {static_cast<long int>(std::chrono::time_point_cast<std::chrono::seconds>(_total).time_since_epoch().count())};
 
     return seconds_to_string(diff);
   }
